refactor: extract exact-CPU trunk runtime (P1)
Moves the exact-CPU decode machinery out of the engine monolith into
Sources/RealModelInference/ExactCPULlamaRuntime.swift:

- CPUExactLlamaRuntime + checkpoint/rollback + forwardToken math
- ExactCPULlamaLayerWeights / CachedExactCPULlamaWeights
- llama weight value types (LlamaQKNormWeights / QKVBias / CPUQKV)
- loadExactCPULlamaLayerWeights, exactCPULlamaLayerForward,
  applyPerHeadRMSNormInPlace, roundFloat16Vector,
  decodeContextFromCaches via a frozen engine extension

Shared kernels (rmsNorm, projections, silu, partitionedArgmax, ...) stay
in the engine and are widened private->internal; no behavior change.

Verified: golden replay bit-identical on all four trunks; full suite green.
